App: Define our own logging module
This disables the built-in Apache-style Rack CommonLogger and uses
our own logger that logs a bunch of information including POST
parameters, timing info, etc.  The format is based on how I have
Rails configured with lograge.

Also setup request.uuid and log it in an X-Request-Id response
header like Rails, which is useful for diagnosing API requests.