Time: Add iso8601_with_ms
The Mastodon iOS app tries to parse a iso8601 string and fails
because it doesn't have milliseconds, and then falls back to just
parsing it as a date.  This screws up the sorting of statuses and
makes everything posted recently "23 hours ago".

Add this method to force the addition of 0 milliseconds so
Mastodon.API.decoder properly uses
fractionalSecondsPreciseISO8601Formatter.